+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Rojo Changelog
 
 ## [Unreleased]
+* Fixed in-memory filesystem not handling out-of-order filesystem change events
 
 ## [0.5.0 Alpha 1](https://github.com/LPGhatguy/rojo/releases/tag/v0.5.0-alpha.1) (January 14, 2019)
 * Changed plugin UI to be way prettier